backend/requirements.txt
djangorestframework
django
markdown
django-filter
drf-nested-routers